Converting code reviews to bulk inserts
This fixes #33. Wohoo!

Now we have a bulk insert running on code reviews. Every model gets
cached in an array. When that array gets too big, it dumps to the
database. To accomplish this, we needed the following

* Associations no longer use their original keys. Fixes #45
* In situations where we need composite keys, I went with a
* composite_ convention. Added new keys and unique indexes for them.
* Some fields needed wider integers, so the database migrated for that
* Removed some problematic properties that we weren't using anyway
* Fixed a verify along the way
* Added more verifys for testing the new associations
* Updated some verify helpers

My benchmarks puts the time saves at 33% of the original time on the
test data set.