Review owners are participants too
This will have some big query implications, but I think the concept is
important. If you initiate a code review, then you are a participant
regardless of whether you send a message.

This was motivated by the small numbers we see in the adjacency_list
table in production. There are lots of code reviews where one person
gives feedback and then the owner just changes the code. That's still
participation.